dx1238: rename entities again
