+++ web-client/integration-tests/journey/petitionsClerkVerifiesOrderDesignatingPlaceOfTrialCheckbox.ts
@@ -1,4 +1,4 @@
-import { IncompletePaperCase } from '../../../shared/src/business/entities/cases/IncompletePaperCase';
+import { IncompletePaperCase } from '../../../shared/src/business/entities/cases/IncompletePaperPetition';
 
 export const petitionsClerkVerifiesOrderDesignatingPlaceOfTrialCheckbox = (
   cerebralTest,
