+++ web-client/integration-tests/journey/petitionsClerkVerifiesPetitionPaymentFeeOptions.ts
@@ -1,6 +1,6 @@
 import { Case } from '../../../shared/src/business/entities/cases/Case';
 import { FORMATS } from '@shared/business/utilities/DateHandler';
-import { IncompletePaperCase } from '../../../shared/src/business/entities/cases/IncompletePaperCase';
+import { IncompletePaperCase } from '../../../shared/src/business/entities/cases/IncompletePaperPetition';
 import { PAYMENT_STATUS } from '../../../shared/src/business/entities/EntityConstants';
 
 export const petitionsClerkVerifiesPetitionPaymentFeeOptions = (
